Create session cookie and accompanying DB
I keep two separate "session" concepts now, one is for the
authentication session which ends pretty quickly, it lasts just long
enough to send a challenge to the browser and get a response back using
the webauthn credentials.

The second session is a normal session cookie, and just identifies a
user after they've authenticated. I wanted to include this in the sample
so that it would be clear to people reading the code that there is a
need to have both sessions to do the usual work of hosting a service.